+++ libraries/resource_providers/api300/synergy/san_manager_provider.rb
@@ -1,4 +1,4 @@
-# (c) Copyright 2016 Hewlett Packard Enterprise Development LP
+# (c) Copyright 2017 Hewlett Packard Enterprise Development LP
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -9,14 +9,14 @@
 # CONDITIONS OF ANY KIND, either express or implied. See the License for the
 # specific language governing permissions and limitations under the License.
 
-OneviewCookbook::ResourceBaseProperties.load(self)
+require_relative '../../api200/san_manager_provider'
 
-default_action :edit
-
-action :edit do
-  OneviewCookbook::Helper.do_resource_action(self, :ServerHardwareType, :edit)
-end
-
-action :remove do
-  OneviewCookbook::Helper.do_resource_action(self, :ServerHardwareType, :remove)
+module OneviewCookbook
+  module API300
+    module Synergy
+      # SANManager API300 Synergy resource provider methods
+      class SANManagerProvider < API200::SANManagerProvider
+      end
+    end
+  end
 end